error with empty line. renamed assembler2 and deleted og
