+++ programs/exampAssemblyCode.txt
@@ -1,1 +1,4 @@
-shift r1 -3
+add2 r2
+sub r3, r1, r1
+//
+xor r2,r2,r2
