+++ programs/output.txt
@@ -1,1 +1,3 @@
-011011101
+101110110
+100011101
+110101010
